+++ R/tar_config_set.R
@@ -86,6 +86,10 @@
 #'   with fast read/write access.
 #'   If the argument `NULL`, the setting is not modified.
 #'   Use [tar_config_unset()] to delete a setting.
+#' @param use_crew Logical of length 1, whether to use `crew` in [tar_make()]
+#'   if the `controller` option is set in `tar_option_set()` in the target
+#'   script (`_targets.R`). See <https://books.ropensci.org/targets/crew.html>
+#'   for details.
 #' @param workers Positive numeric of length 1, `workers` argument of
 #'   [tar_make_clustermq()] and related functions that run the pipeline
 #'   with parallel computing among targets.
@@ -143,6 +147,7 @@ tar_config_set <- function(
   seconds_interval = NULL,
   store = NULL,
   shortcut = NULL,
+  use_crew = NULL,
   workers = NULL,
   config = Sys.getenv("TAR_CONFIG", "_targets.yaml"),
   project = Sys.getenv("TAR_PROJECT", "main")
@@ -163,6 +168,7 @@ tar_config_set <- function(
   tar_config_assert_seconds_interval(seconds_interval)
   tar_config_assert_shortcut(shortcut)
   tar_config_assert_store(store)
+  tar_config_assert_use_crew(use_crew)
   tar_config_assert_workers(workers)
   yaml <- tar_config_read_yaml(config)
   if (!tar_config_is_multi_project(yaml, config)) {
@@ -183,6 +189,7 @@ tar_config_set <- function(
     yaml[[project]]$seconds_interval
   yaml[[project]]$shortcut <- shortcut %|||% yaml[[project]]$shortcut
   yaml[[project]]$store <- store %|||% yaml[[project]]$store
+  yaml[[project]]$use_crew <- use_crew %|||% yaml[[project]]$use_crew
   yaml[[project]]$workers <- if_any(
     is.null(workers),
     yaml[[project]]$workers,
@@ -280,6 +287,15 @@ tar_config_assert_store <- function(store) {
   }
   tar_assert_scalar(store)
   tar_assert_chr(store)
+}
+
+tar_config_assert_use_crew <- function(use_crew) {
+  if (is.null(use_crew)) {
+    return()
+  }
+  tar_assert_lgl(use_crew)
+  tar_assert_scalar(use_crew)
+  tar_assert_none_na(use_crew)
 }
 
 tar_config_assert_workers <- function(workers) {