+++ inst/pipelines/use_targets.R
@@ -9,12 +9,28 @@ library(targets)
 
 # Set target options:
 tar_option_set(
-  packages = c("tibble"), # packages that your targets need to run
-  format = "rds" # default storage format
+  packages = c("tibble") # packages that your targets need to run
+  # format = "qs", # Optionally set the default storage format. qs is fast.
   #
   # For distributed computing in tar_make(), supply a {crew} controller
-  #   as discussed at https://books.ropensci.org/targets/crew.html
-  # controller = crew::crew_controller_local(workers = 2)
+  # as discussed at https://books.ropensci.org/targets/crew.html.
+  # Choose a controller that suits your needs. For example, the following
+  # sets a controller with 2 workers which will run as local R processes:
+  #
+  #   controller = crew::crew_controller_local(workers = 2)
+  #
+  # Alternatively, if you want workers to run on a high-performance computing
+  # cluster, select a controller from the {crew.cluster} package. The following
+  # example is a controller for Sun Grid Engine (SGE).
+  # 
+  #   controller = crew.cluster::crew_controller_sge(
+  #     workers = 50,
+  #     # Many clusters install R as an environment module, and you can load it
+  #     # with the script_lines argument. To select a specific verison of R,
+  #     # you may need to include a version string, e.g. "module load R/4.3.0".
+  #     # Check with your system administrator if you are unsure.
+  #     script_lines = "module load R"
+  #   )
   #
   # Set other options as needed.
 )